Fixed a minor bug related to divide by zero
There was an exception made so that the compiler would not raise a divide by zero error if and only if the first operand was a zero-float, and the second operand was also a zero-float. This is the only way to get NAN.

There was a memory access issue that was now fixd.